+++ lib/CMakeLists.txt
@@ -2,10 +2,8 @@ add_subdirectory(CAPI)
 add_subdirectory(Conversion)
 add_subdirectory(Dialect)
 add_subdirectory(RefBackend)
-add_subdirectory(JITRuntime)
 add_subdirectory(Python)
 add_subdirectory(Typing)
-add_subdirectory(runtime)
 
 if(NPCOMP_ENABLE_REFJIT)
   add_subdirectory(Backend/RefJIT)
