+++ lib/RefBackend/JITHelpers/JITModule.cpp
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "npcomp/JITRuntime/JITModule.h"
+#include "npcomp/RefBackend/JITHelpers/JITModule.h"
 #include "mlir/ExecutionEngine/CRunnerUtils.h"
 #include "mlir/ExecutionEngine/OptUtils.h"
 #include "npcomp/RefBackend/RefBackend.h"
